modificate relazioni ed entità follow e user
