+++ src/main/java/co/develhope/spring/entities/ILike.java
@@ -1,0 +1,29 @@
+package co.develhope.spring.entities;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Table
+public class ILike {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "UserId")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "CommentId")
+    private Comment comment;
+    private LocalDateTime dateTime;
+}
