+++ src/main/java/co/develhope/spring/repositories/FollowRepository.java
@@ -1,7 +1,7 @@
-package com.develhope.spring.repositories;
+package co.develhope.spring.repositories;
 
-import com.develhope.spring.entities.Follow;
-import com.develhope.spring.entities.User;
+import co.develhope.spring.entities.Follow;
+import co.develhope.spring.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +10,4 @@ import java.util.List;
 @Repository
 public interface FollowRepository extends JpaRepository<Follow, Long> {
     List<Follow> findByFollower(User follower);
-    List<Follow> findByFollowed(User followed);
-    Follow findByFollowerAndFollowed(User follower, User followed);
 }
