+++ src/main/java/co/develhope/spring/repositories/ILikeRepository.java
@@ -1,6 +1,6 @@
-package com.develhope.spring.repositories;
+package co.develhope.spring.repositories;
 
-import com.develhope.spring.entities.ILike;
+import co.develhope.spring.entities.ILike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
